Refactor and add code for automatically updating the calendar and using slack as an event feed.

+++ show-time.py
@@ -1,4 +1,4 @@
 import datetime
 import sys
 
-print(datetime.datetime.fromtimestamp(int(sys.argv[1])).strftime('%c'))
+print(datetime.datetime.fromtimestamp(float(sys.argv[1])).strftime('%c'))
